refactor to move image handling to routes and controller
