+++ question-service/routes/routes.js
@@ -3,6 +3,7 @@ import { getAllQuestions, getQuestionById, getQuestionByDifficulty, getQuestionB
 import { createNewQuestion } from '../controller/create.js';
 import { updateQuestion } from '../controller/update.js';
 import { deleteQuestion } from '../controller/delete.js';
+import { getImage, uploadImage } from "../controller/imageController.js";
 
 const router = Router();
 
@@ -43,5 +44,12 @@ router.put('/:id', updateQuestion);
 
 router.delete('/:id', deleteQuestion);
 
+/**
+ * IMAGE HANDLING
+ */
+
+router.get('/img/:filename', getImage);
+
+router.post('/img', uploadImage);
 
 export default router;
